Converting dataset from drive to PASCAL
Defines a class that can be used to convert the uploaded annotations done on drive to a PASCAL formatted dataset that is acceptable to most ML algorithms. We will use this to create our datasets. There are components of the methods currently that require manually input (when there were mistakes made by annotators). So this data generate pipeline in its current implementation is run from a notebook